Payload_Type/thanatos/requirements.txt
dynaconf==3.1.4
aio-pika
mythic-payloadtype-container==0.1.1
donut-shellcode==0.9.2
